+++ lib/papierkram_api/v1/endpoints/projects.rb
@@ -55,10 +55,39 @@ module PapierkramApi
           http_post("#{@url_api_path}/projects", body)
         end
 
-        def update_by(id:, attributes: {})
-          raise ArgumentError, 'attributes must be a Hash' unless attributes.is_a?(Hash)
+        def update_by( # rubocop:disable Metrics/AbcSize, Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity, Metrics/ParameterLists
+          id:,
+          name: nil,
+          customer_id: nil,
+          description: nil,
+          start_date: nil,
+          end_date: nil,
+          flagged: nil,
+          budget_type: nil,
+          budget_money: nil,
+          budget_time: nil,
+          budget_time_unit: nil,
+          color: nil,
+          default_proposition: nil,
+          team_members: nil
+        )
 
-          http_put("#{@url_api_path}/projects/#{id}", attributes)
+          body = {}
+          body[:name] = name if name
+          body[:customer] = { id: customer_id } if customer_id
+          body[:description] = description if description
+          body[:start_date] = start_date if start_date
+          body[:end_date] = end_date if end_date
+          body[:flagged] = flagged if flagged
+          body[:budget_type] = budget_type if budget_type
+          body[:budget_money] = budget_money if budget_money
+          body[:budget_time] = budget_time if budget_time
+          body[:budget_time_unit] = budget_time_unit if budget_time_unit
+          body[:color] = color if color
+          body[:default_proposition] = default_proposition if default_proposition
+          body[:team_members] = team_members if team_members
+
+          http_put("#{@url_api_path}/projects/#{id}", body)
         end
 
         def delete_by(id:)